pyproject.toml
[tool.poetry]
name = "conspectium-backend"
version = "0.1.0"
description = "Backend service for Conspectium MAX Mini App"
authors = ["Conspectium Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.3.0"
alembic = "^1.13.1"
python-multipart = "^0.0.9"
google-generativeai = "^0.5.4"
pyjwt = "^2.8.0"
httpx = "^0.27.0"
passlib = "^1.7.4"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
faker = "^25.0.0"
ruff = "^0.4.0"
black = "^24.4.2"
mypy = "^1.9.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "ASYNC",
    "S",
    "B",
]
ignore = [
    "S101",
]

[tool.ruff.isort]
known-first-party = ["app"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
